requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
mysqlclient==2.2.8
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
